refactor: Renaming IFO icon to IPO
