+++ src/widgets/Menu/icons/Ipo.tsx
