+++ src/widgets/Menu/icons/index.ts
@@ -3,7 +3,7 @@ export { default as GroupsIcon } from "./Groups";
 export { default as HamburgerIcon } from "./Hamburger";
 export { default as HamburgerCloseIcon } from "./HamburgerClose";
 export { default as HomeIcon } from "./Home";
-export { default as IfoIcon } from "./Ifo";
+export { default as IpoIcon } from "./Ipo";
 export { default as InfoIcon } from "./Info";
 export { default as LanguageIcon } from "./Language";
 export { default as LogoIcon } from "./Logo";
